reverting old changes and loading script once
