+++ aemedge/scripts/scripts.js
@@ -20,7 +20,6 @@ import {
   getPageType,
   buildFragmentBlocks,
   createTag,
-  loadGameFinderApp,
 } from './utils.js';
 
 const LCP_BLOCKS = ['category']; // add your LCP blocks to the list
@@ -400,11 +399,6 @@ async function loadEager(doc) {
     decorateMain(main);
     await loadTemplate(main);
     document.body.classList.add('appear');
-    const gameFinders = doc.querySelectorAll('.game-finder.block');
-    if (gameFinders && gameFinders.length > 0) {
-      await loadGameFinderApp();
-    }
-
     await waitForLCP(LCP_BLOCKS);
   }
