+++ aemedge/scripts/utils.js
@@ -537,23 +537,3 @@ export async function fetchGQL(query, variables, operationName) {
   const gqlResponse = await res.json();
   return gqlResponse;
 }
-
-async function loadScript(src, attrs) {
-  return new Promise((resolve, reject) => {
-    const script = document.createElement('script');
-    script.src = src;
-    if (attrs) {
-      // eslint-disable-next-line no-restricted-syntax, guard-for-in
-      for (const attr in attrs) {
-        script.setAttribute(attr, attrs[attr]);
-      }
-    }
-    script.onload = resolve;
-    script.onerror = reject;
-    document.querySelector('body').append(script);
-  });
-}
-
-export async function loadGameFinderApp() {
-  loadScript('/aemedge/scripts/sling-react/gamefinder-build.js');
-}
